Commit
Slight edits to the HTML and created an object class for the Restaurants. The idea being that they can be churned out once input can be given by a database or JSON file. 

Restaurants will be objects that will have strings for name and description and then sent over through url_for() with flask on the front end. Some other data like directions, stars on yelp, phone number, and emails can eventually be added.

Thus far, MySQL is the intended database technology that will be used because I have used it before and feel comfortable using it. I believe that the simplicity of the project will lend itself well to using MySQL.

to do:

1. Object Creation
2. Object passing to the html
3. Database creation
4. Additional features.
5. Refinement.